Thu, Mar 2, 2023, 9:17 AM +08:00

+++ hours/hours.c
@@ -27,10 +27,10 @@ int main(void)
 // TODO: complete the calc_hours function
 float calc_hours(int hours[], int weeks, char output)
 {
-    float total;
-    int average;
+    int total;
+    float average;
     int i;
-    for (i = 0, total = 0.0; i < weeks; i++) //add hours into a variable
+    for (i = 0, total = 0; i < weeks; i++) //add hours into a variable
     {
         total += hours[i];
     }
